+++ src/main/java/com/parceiroferramentas/api/parceiro_api/repository/FerramentaRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import com.parceiroferramentas.api.parceiro_api.model.Ferramenta;
 
 
-public interface FerramentaRepository extends JpaRepository<Ferramenta, Integer> {
+public interface FerramentaRepository extends JpaRepository<Ferramenta, Long> {
 
     public Page<Ferramenta> findByTipoEqualsIgnoreCase(String tipo, Pageable pageable);
 }
